Corrections in the tests
