+++ src/einsteinpy/tests/test_metric/test_kerrnewman.py
@@ -8,12 +8,7 @@ from numpy.testing import assert_allclose
 from einsteinpy import constant
 from einsteinpy.coordinates import BoyerLindquistDifferential, CartesianDifferential
 from einsteinpy.metric import KerrNewman
-from einsteinpy.utils import (
-    kerr_utils,
-    kerrnewman_utils,
-    schwarzschild_radius,
-    schwarzschild_radius_dimensionless,
-)
+from einsteinpy.utils import kerrnewman_utils, schwarzschild_radius
 
 _c = constant.c.value
 _G = constant.G.value
@@ -167,25 +162,3 @@ def test_calculate_trajectory_iterator_RuntimeWarning():
         for _, _ in zip(range(1000), it):
             pass
         assert len(w) >= 1
-
-
-def test_event_horizon_for_nonrotating_uncharged_case():
-    M = 5e27
-    a = 0.0
-    Q = 0.0
-    _scr = schwarzschild_radius_dimensionless(M)
-    a1 = kerrnewman_utils.event_horizon(M, a, Q, np.pi / 4, "Spherical")
-    a2 = kerr_utils.event_horizon(M, a, np.pi / 4)
-    assert_allclose(a1, a2, rtol=0.0, atol=1e-4)
-    assert_allclose(a1[0], _scr, rtol=0.0, atol=1e-4)
-
-
-def test_radius_ergosphere_for_nonrotating_uncharged_case():
-    M = 5e27
-    a = 0.0
-    Q = 0.0
-    _scr = schwarzschild_radius_dimensionless(M)
-    a1 = kerrnewman_utils.radius_ergosphere(M, a, Q, np.pi / 4, "Spherical")
-    a2 = kerr_utils.radius_ergosphere(M, a, np.pi / 4)
-    assert_allclose(a1, a2, rtol=0.0, atol=1e-4)
-    assert_allclose(a1[0], _scr, rtol=0.0, atol=1e-4)
